+++ packages/dataviews/src/dataform-controls/index.tsx
@@ -21,8 +21,8 @@ import integer from './integer';
 import radio from './radio';
 import select from './select';
 import text from './text';
+import toggle from './toggle';
 import toggleGroup from './toggle-group';
-import boolean from './boolean';
 import array from './array';
 
 interface FormControls {
@@ -31,7 +31,6 @@ interface FormControls {
 
 const FORM_CONTROLS: FormControls = {
 	array,
-	boolean,
 	checkbox,
 	datetime,
 	date,
@@ -42,6 +41,7 @@ const FORM_CONTROLS: FormControls = {
 	radio,
 	select,
 	text,
+	toggle,
 	toggleGroup,
 };